requirements.txt
discord.py>=2.3.0
groq>=0.9.0
